updated some stufff
