+++ backend/src/models/models.ts
@@ -3,8 +3,7 @@ export class Subscription {
     public name: string,
     public image: string,
     public cost: number,
-    public billing_date: string,
-    public next_billing_date: string
+    public billing_date: string
   ) {}
 }
 export class Group {
